+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 ### Big Data Types v0.5.1
-- Spark: Added methods to convert Instances into Spark inside `SparkSchemas`.
-- Spark: New extension methods that will allow any other instance to converted into Spark types
+- Spark: 
+  - Added a missing implementation for derive instances into SparkSchemas
+  - Added methods to convert Instances into Spark inside `SparkSchemas`.
+  - New extension methods that will allow any other instance to converted into Spark types
 - BigQuery: Fixed some names in extension methods
 - New cross module examples in Examples module
+  - Added examples of conversions from BigQuery to Spark and Cassandra
 
 ### Big Data Types v0.5.0
 - BigQuery: Reverse conversion.